fixed tail hang on file removal
Previously `tail -f` was used for rules that followed files.
But if the file was moved, such as during a log rotate,
the daemon would silently stop getting data since it has
a file handle to a file which has been removed.
To fix this the command has been changed to
`tail --follow=name --retry` and it now behaves correctly.